dcc.c: Report errno string when file open() for writing fails. outbound/text/textgui: Remove pointer casts to (long). servlist.c: Update Aitvaras network. text.c: Add feature to auto update old text events. textevents.in: Add $2 arg for "DCC RECV File Open Error".
git-svn-id: https://xchat.svn.sourceforge.net/svnroot/xchat@507 893a96be-7f27-4fdf-9d1e-6aeec9d3cce1

+++ src/common/outbound.c
@@ -3054,13 +3054,13 @@ nick_comp_cb (struct User *user, nickdata *data)
 static void
 perform_nick_completion (struct session *sess, char *cmd, char *tbuf)
 {
-	long len;
+	int len;
 	char *space = strchr (cmd, ' ');
 	if (space && space != cmd)
 	{
 		if (((space[-1] == ':') || (space[-1] == prefs.nick_suffix[0])) && (space - 1 != cmd))
 		{
-			len = (long) space - (long) cmd - 1;
+			len = space - cmd - 1;
 			if (len < NICKLEN)
 			{
 				char nick[NICKLEN];
